+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/CloseLocationValueIndicator.java
@@ -36,6 +36,8 @@ import org.ta4j.core.num.Num;
  */
 public class CloseLocationValueIndicator extends CachedIndicator<Num> {
 
+    private final Num zero = numOf(0);
+
     public CloseLocationValueIndicator(TimeSeries series) {
         super(series);
     }
@@ -43,11 +45,12 @@ public class CloseLocationValueIndicator extends CachedIndicator<Num> {
     @Override
     protected Num calculate(int index) {
         final Bar bar = getTimeSeries().getBar(index);
-
-        final Num close = bar.getClosePrice();
         final Num low = bar.getLowPrice();
         final Num high = bar.getHighPrice();
+        final Num close = bar.getClosePrice();
 
-        return ((close.minus(low)).minus(high.minus(close))).dividedBy(high.minus(low));
+        final Num diffHighLow = high.minus(low);
+
+        return diffHighLow.isNaN() ? zero : ((close.minus(low)).minus(high.minus(close))).dividedBy(diffHighLow);
     }
 }