+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **StopLossRule**: now correctly handles stops for sell orders
 - **ProfitLossCriterion**: fixed to work properly for short trades
 - **PivotPointIndicator**: fixed possible npe if first bar is not in same period
+- **CloseLocationValueIndicator**: fixed special case, return zero instead of NaN if high price == low price
 
 ### Changed
 - **BaseTimeSeriesBuilder** moved from inner class to own class